requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
ultralytics==8.3.198
opencv-python-headless==4.11.0.86
pillow==10.2.0
python-multipart==0.0.20
numpy==1.26.4
pydantic== 2.11.7
